Move plugins v1 code into an isolated directory
