+++ lib/inspec/plugin/v1.rb
@@ -1,0 +1,2 @@
+require 'inspec/plugin/v1/plugins'
+require 'inspec/plugin/v1/registry'
