+++ lib/inspec/plugin/v1/plugin_types/resource.rb
