+++ lib/inspec/plugin/v1/plugin_types/source_reader.rb
@@ -2,7 +2,7 @@
 # author: Dominik Richter
 # author: Christoph Hartmann
 
-require 'utils/plugin_registry'
+require 'inspec/plugin/v1/registry'
 
 module Inspec
   module Plugins
