+++ lib/inspec/plugin/v1/registry.rb
